feat(GenericContainer): add WorkdirSetting trait for managing container working directory
